// packages/api-client/src/generated/core.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export type paths = Record<string, never>;
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        Error: {
            code: string;
            message: string;
            details?: {
                [key: string]: unknown;
            };
        };
        Pagination: {
            page: number;
            pageSize: number;
            total: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;